% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABCSMC.R
\name{ABC_SMC}
\alias{ABC_SMC}
\title{Runs ABC-SMC algorithm}
\usage{
ABC_SMC(npart, tols, priors, func, data)
}
\arguments{
\item{npart}{The number of particles (must be a positive integer).}

\item{tols}{A \code{matrix} of tolerances, with the number of rows defining
the number of generations of required, and the number of columns
defining the number of summary statistics / data points to match to.}

\item{priors}{A \code{matrix} containing the lower and upper bounds of uniform
priors; with number of rows equal to the number of parameters, and
two columns containing the lower and upper bounds respectively.}

\item{func}{A function taking a single argument \code{pars} that runs the simulator
and returns the simulated summary measures against which to compare.}

\item{data}{A \code{vector} containing the observed summary statistics to match to.}
}
\value{
An \code{ABCSMC} object, essentially a \code{list} containing:
\itemize{
\item{\code{pars}:}{ a \code{list} of \code{matrix} objects containing the accepted
                     particles. Each element of the list corresponds to a generation 
                     of ABC-SMC, with each matrix being of dimension 
                     \code{npart} x \code{npars};}
\item{\code{output}:}{ a \code{list} of \code{matrix} objects containing the simulated
                     summary statistics. Each element of the list corresponds to a
                     generation of ABC-SMC, with each matrix being of dimension 
                     \code{npart} x \code{length(data)};}
\item{\code{weights}:}{ a \code{list} of \code{vector} objects containing the particle
                     weights. Each element of the list corresponds to a
                     generation of ABC-SMC, with each vector being of length
                     \code{npart};}
\item{\code{accrate}:}{ a \code{vector} of length \code{nrow(tols)} containing the
                     acceptance rates for each generation of ABC.}
}
}
\description{
Runs ABC-SMC algorithm of Toni et al. (2009)
}
