% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABCRef.R
\name{ABCRef}
\alias{ABCRef}
\title{Produces ABC reference table}
\usage{
ABCRef(npart, priors, func, data, parallel = F, mc.cores = NA, ...)
}
\arguments{
\item{npart}{The number of particles (must be a positive integer).}

\item{priors}{A \code{data.frame} containing columns \code{parnames}, \code{dist}, \code{p1} and 
\code{p2}, with number of rows equal to the number of parameters. The column
\code{parname} simply gives names to each parameter for plotting and summarising.
Each entry in the \code{dist} column must contain one of \code{c("unif", "norm", "gamma")}, 
and the corresponding \code{p1} and \code{p2} entries relate to the hyperparameters 
(lower and upper bounds in the uniform case; mean and standard deviation in the 
normal case; and shape and rate in the gamma case).}

\item{func}{A function takes a single argument \code{pars}. The function runs the simulator returns a 
\code{vector} of simulated summary measures or a missing value (\code{NA}) if there is an 
error in the simulator.}

\item{data}{A \code{data.frame} with two columns containing the observed summary statistics. 
The first column must be called \code{outnames} and contain the output names, and the
second column must be called \code{values} and contain the observations.}

\item{parallel}{A \code{logical} determining whether to use parallel processing or not.}

\item{mc.cores}{Number of cores to use if using parallel processing.}
}
\value{
An \code{data.frame} object with \code{npart} rows, where the first \code{p} columns correspond to 
        the proposed parameters, and the remaining columns correspond to the simulated outputs.
}
\description{
Runs ABC reference table
}
