% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABCSMC.R
\name{ABCSMC}
\alias{ABCSMC}
\alias{ABCSMC.ABCSMC}
\alias{ABCSMC.default}
\title{Runs ABC-SMC algorithm}
\usage{
ABCSMC(x, ...)

\method{ABCSMC}{ABCSMC}(x, tols, parallel = F, mc.cores = NA)

\method{ABCSMC}{default}(x, data, priors, func, tols, parallel = F,
  mc.cores = NA, ...)
}
\arguments{
\item{x}{An \code{ABCSMC} object or the number of particles (must be a positive integer).}

\item{...}{Further arguments to pass to \code{func}. (Not used if extending runs.)}

\item{tols}{A \code{data.frame} of tolerances, with the number of rows defining
the number of generations required, and columns defining the summary statistics
to match to. The columns must match to those in `data`.}

\item{parallel}{A \code{logical} determining whether to use parallel processing or not.}

\item{mc.cores}{Number of cores to use if using parallel processing.}

\item{data}{A \code{data.frame} with a single row and columns containing the observed summary statistics
to match to. Columns must match to `tols`.}

\item{priors}{A \code{data.frame} containing columns \code{parnames}, \code{dist}, \code{p1} and 
\code{p2}, with number of rows equal to the number of parameters. The column
\code{parname} simply gives names to each parameter for plotting and summarising.
Each entry in the \code{dist} column must contain one of \code{c("unif", "norm", "gamma")}, 
and the corresponding \code{p1} and \code{p2} entries relate to the hyperparameters 
(lower and upper bounds in the uniform case; mean and standard deviation in the 
normal case; and shape and rate in the gamma case).}

\item{func}{Function that runs the simulator and checks whether the simulation matches the data. 
The first four arguments must be \code{pars}, \code{data}, \code{tols} and 
\code{iniStates}. If the simulations do not match the data then the function must 
return an \code{NA}, else it must returns a \code{vector} of simulated summary measures. 
In this latter case the output from the function must be a vector with length equal to 
\code{ncol(data)} and with entries in the same order as the columns of \code{data}.}
}
\value{
An \code{ABCSMC} object, essentially a \code{list} containing:
\itemize{
\item{\code{pars}:}{ a \code{list} of \code{matrix} objects containing the accepted
                     particles. Each element of the list corresponds to a generation 
                     of ABC-SMC, with each matrix being of dimension 
                     \code{npart} x \code{npars};}
\item{\code{output}:}{ a \code{list} of \code{matrix} objects containing the simulated
                     summary statistics. Each element of the list corresponds to a
                     generation of ABC-SMC, with each matrix being of dimension 
                     \code{npart} x \code{ncol(data)};}
\item{\code{weights}:}{ a \code{list} of \code{vector} objects containing the particle
                     weights. Each element of the list corresponds to a
                     generation of ABC-SMC, with each vector being of length
                     \code{npart};}
\item{\code{accrate}:}{ a \code{vector} of length \code{nrow(tols)} containing the
                     acceptance rates for each generation of ABC;}
\item{\code{tols}:}{ a copy of the \code{tols} input;}
\item{\code{priors}:}{ a copy of the \code{priors} input;}
\item{\code{data}:}{ a copy of the \code{data} input;}
\item{\code{func}:}{ a copy of the \code{func} input}
\item{\code{addargs}:}{ a copy of the \code{...} inputs.}
}
}
\description{
Runs ABC-SMC algorithm of Toni et al. (2009) for fitting 
                 infectious disease models to time series count data
}
