% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileRcpp.R
\name{compileRcpp}
\alias{compileRcpp}
\title{Compile parsed Rcpp model}
\usage{
compileRcpp(parsed = NULL)
}
\arguments{
\item{parsed:}{An object of class \code{parsedRcpp}.}
}
\value{
An object of class \code{XPtr} that points to the compiled function, or
        an R \code{function} object for calling directly from R.
}
\description{
Compiles an object of class \code{parsedRcpp} into an
             \code{XPtr} object for use in Rcpp functions, or an
             object of class \code{function} for calling directly from R.
}
