% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ebola}
\alias{ebola}
\title{Time series counts of ebola cases}
\format{
A data frame with 192 rows and 3 variables:
\describe{
  \item{time}{days from 1st January 1995}
  \item{clin_signs}{number of new clinical cases at each day}
  \item{removals}{number of new removals at each day}
}
}
\source{
Khan AS et al. (1999) <doi:10.1086/514306>
}
\usage{
ebola
}
\description{
A dataset containing time series counts for the number
     of new individuals exhibiting clinical signs, and the number of
     new removals each day for the 1995 Ebola epidemic in the Democratic
     Republic of Congo
}
\keyword{datasets}
