% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileRcpp.R, R/mparse_Rcpp.R
\name{mparse_Rcpp}
\alias{mparse_Rcpp}
\title{Compile parsed Rcpp model}
\usage{
mparse_Rcpp(transitions = NULL, compartments = NULL, pars = NULL)

mparse_Rcpp(transitions = NULL, compartments = NULL, pars = NULL)
}
\arguments{
\item{parsed:}{An object of class \code{parsedRcpp}.}

\item{transitions:}{character vector containing transitions on the form \code{"X ->
... -> Y"}. The left (right) side is the initial (final)
state and the propensity is written in between the
\code{->}-signs. The special symbol \code{@} is reserved for the empty
set. For example, \code{transitions = c("S -> k1*S*I -> I", "I ->
k2*I -> R")} expresses a SIR model.}

\item{compartments:}{contains the names of the involved compartments, for
example, \code{compartments = c("S", "I", "R")}.}

\item{pars:}{named vector of parameters.}
}
\value{
An object of class \code{XPtr} that points to the compiled function.

An object of class \code{parsedRcpp} that contains code to compile
        into an \code{XPtr} object.
}
\description{
Compiles an object of class \code{parsedRcpp} into an
             \code{XPtr} object for use in \code{ABCSMC} functions.

Parse custom model using \code{SimInf} style markup and 
create code to compile to an \code{XPtr} object to use in \code{ABCSMC} functions.
Does not have full functionality of \code{mparse}. Currently only supports
simulations on a single node.
}
