% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPMCMC.R
\name{plot.PMCMC}
\alias{plot.PMCMC}
\title{Plots \code{PMCMC} objects}
\usage{
\method{plot}{PMCMC}(x, type = c("post", "trace"), joint = F,
  transfunc = NA, ask = T)
}
\arguments{
\item{x}{A \code{PMCMC} object.}

\item{type}{Takes the value \code{"post"} if you want to plot posterior distributions.
Takes the value \code{"trace"} if you want to plot the trace plots.}

\item{joint}{A logical describing whether joint or marginal distributions are wanted.}

\item{transfunc}{Is a \code{function} object where the arguments to the function must
match all or a subset of the parameters in the model. This function needs 
to return a \code{data.frame} object with columns containing the transformed
parameters.}

\item{ask}{Should the user ask before moving onto next trace plot.}
}
\value{
A plot of the (approximate) posterior distributions from the particle MCMC algorithm,
        or corresponding trace plots.
}
\description{
Plot method for \code{PMCMC} objects.
}
