% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSimBIID_runs.R
\name{plot.SimBIID_runs}
\alias{plot.SimBIID_runs}
\title{Plots \code{SimBIID_runs} objects}
\usage{
plot.SimBIID_runs(x, which = c("all", "t"), type = c("runs", "sums"),
  rep = NA, quant = 0.9, ...)
}
\arguments{
\item{x}{An \code{SimBIID_runs} object.}

\item{which}{A character vector of states to plot. Can be \code{"all"} to plot all
states (and final event times), or \code{"t"} to plot final event times.}

\item{type}{Character stating whether to plot full simulations over time (\code{"runs"}) or
summaries (\code{"sums"}).}

\item{rep}{An integer vector of simulation runs to plot.}

\item{quant}{A vector of quantiles (> 0.5) to plot if \code{type == "runs"}.}

\item{...}{Not used here.}
}
\value{
A plot of individual simulations and/or summaries of repeated simulations 
        extracted from \code{SimBIID_runs} object.
}
\description{
Plot method for \code{SimBIID_runs} objects.
}
