% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictPMCMC.R
\name{predict.PMCMC}
\alias{predict.PMCMC}
\title{Predicts future course of outbreak from \code{PMCMC} objects}
\usage{
\method{predict}{PMCMC}(object, tspan, npart = 50, ...)
}
\arguments{
\item{object}{A \code{PMCMC} object.}

\item{tspan}{A vector of times over which to output predictions.}

\item{npart}{The number of particles to use in the bootstrap filter.}

\item{...}{Not used here.}
}
\value{
A \code{SimBIID_runs} object.
}
\description{
Predict method for \code{PMCMC} objects.
}
