% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printABC.R
\name{print.ABCSMC}
\alias{print.ABCSMC}
\title{Prints \code{ABCSMC} objects}
\usage{
\method{print}{ABCSMC}(x, ...)
}
\arguments{
\item{x}{An \code{ABCSMC} object.}

\item{...}{Not used here.}
}
\value{
Summary outputs printed to the screen.
}
\description{
Print method for \code{ABCSMC} objects.
}
