% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printPMCMC.R
\name{print.PMCMC}
\alias{print.PMCMC}
\title{Prints \code{PMCMC} objects}
\usage{
\method{print}{PMCMC}(x, ...)
}
\arguments{
\item{x}{A \code{PMCMC} object.}

\item{...}{Not used here.}
}
\value{
Summary outputs printed to the screen.
}
\description{
Print method for \code{PMCMC} objects.
}
