% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printSimBIID_model.R
\name{print.SimBIID_model}
\alias{print.SimBIID_model}
\title{Prints \code{SimBIID_model} objects}
\usage{
\method{print}{SimBIID_model}(x, ...)
}
\arguments{
\item{x}{A \code{SimBIID_model} object.}

\item{...}{Not used here.}
}
\value{
Prints parsed \code{Rcpp} code to the screen.
}
\description{
Print method for \code{SimBIID_model} objects.
}
