% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printSimBIID_runs.R
\name{print.SimBIID_runs}
\alias{print.SimBIID_runs}
\title{Prints \code{SimBIID_runs} objects}
\usage{
\method{print}{SimBIID_runs}(x, ...)
}
\arguments{
\item{x}{A \code{SimBIID_runs} object.}

\item{...}{Not used here.}
}
\value{
Summary outputs printed to the screen.
}
\description{
Print method for \code{SimBIID_runs} objects.
}
