% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimBIID-package.R
\docType{package}
\name{SimBIID-package}
\alias{SimBIID-package}
\title{Simulation-based inference for infectious disease models}
\description{
Package implements various simulation-based inference routines for infectious
disease models.
}
\details{
Package provides some code to run simulations of state-space models, and then
use these in the ABC-SMC algorithm of Toni et al. (2009)
and the bootstrap particle filter based particle MCMC algorithm (Andrieu et al., 2010). 
Also provides functions to plot and summarise the outputs.
}
\author{
TJ McKinley <t.mckinley@exeter.ac.uk>
}
