% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimBIID-package.R
\docType{package}
\name{SimBIID}
\alias{SimBIID}
\alias{SimBIID-package}
\title{Simulation-based inference for infectious disease models}
\description{
Package implements various simulation-based inference routines for infectious
disease models.
}
\details{
Provides some code to run simulations of state-space models, and then
 use these in the Approximate Bayesian Computation Sequential Monte Carlo (ABC-SMC) 
 algorithm of Toni et al. (2009) <doi:10.1098/rsif.2008.0172> and a bootstrap particle
 filter based particle Markov chain Monte Carlo (PMCMC) algorithm 
 (Andrieu et al., 2010 <doi:10.1111/j.1467-9868.2009.00736.x>). 
 Also provides functions to plot and summarise the outputs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tjmckinley/SimBIID}
  \item Report bugs at \url{https://github.com/tjmckinley/SimBIID/issues}
}

}
\author{
\strong{Maintainer}: Trevelyan J. McKinley \email{t.mckinley@exeter.ac.uk}

Authors:
\itemize{
  \item Stefan Widgren \email{stefan.widgren@gmail.com} (Author of 'R/mparse.R')
}

Other contributors:
\itemize{
  \item Pavol Bauer (R/mparse.R) [copyright holder]
  \item Robin Eriksson (R/mparse.R) [copyright holder]
  \item Stefan Engblom (R/mparse.R) [copyright holder]
}

}
