% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimInf_mparse_Rcpp.R
\name{SimInf_mparse_Rcpp}
\alias{SimInf_mparse_Rcpp}
\title{Parse custom model using \code{SimInf} style markup}
\usage{
SimInf_mparse_Rcpp(transitions = NULL, compartments = NULL,
  gdata = NULL)
}
\value{
An object of class \code{XPtr} that points to the compiled function.
}
\description{
Parse custom model using \code{SimInf} style markup and 
create pointer to compiled Rcpp code to use in \code{ABCSMC} functions.
Does not have full functionality of \code{mparse}. Currently only supports
simulations on a single node.
}
