% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryABC.R
\name{summary.ABCSMC}
\alias{summary.ABCSMC}
\title{Summarises \code{ABCSMC} objects}
\usage{
\method{summary}{ABCSMC}(object, gen = NA)
}
\arguments{
\item{object}{An \code{ABCSMC} object.}

\item{gen}{The generation of ABC that you want to extract. If left missing then
defaults to final generation.}
}
\value{
A \code{data.frame} with weighted posterior means and variances.
}
\description{
Summary method for \code{ABCSMC} objects.
}
