% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryPMCMC.R
\name{summary.PMCMC}
\alias{summary.PMCMC}
\title{Summarises \code{PMCMC} objects}
\usage{
\method{summary}{PMCMC}(object, transfunc = NA, ...)
}
\arguments{
\item{object}{A \code{PMCMC} object.}

\item{transfunc}{Is a \code{function} object where the arguments to the function must
match all or a subset of the parameters in the model. This function needs 
to return a \code{data.frame} object with columns containing the transformed
parameters.}

\item{...}{Not used here.}
}
\value{
A \code{summary.mcmc} object.
}
\description{
Summary method for \code{PMCMC} objects.
}
