% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windowPMCMC.R
\name{window.PMCMC}
\alias{window.PMCMC}
\title{Time windows for \code{PMCMC} objects}
\usage{
\method{window}{PMCMC}(x, ...)
}
\arguments{
\item{x}{a \code{PMCMC} object, usually as a result of a call to
\code{PMCMC}.}

\item{\dots}{arguments to pass to \code{\link{window.mcmc}}}
}
\value{
a \code{PMCMC} object
}
\description{
\code{window} method for class \code{PMCMC}
}
\details{
Acts as a wrapper function for \code{\link[coda]{window.mcmc}} 
         from the \code{coda} package
}
