% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABCRef.R
\name{ABCRef}
\alias{ABCRef}
\title{Produces ABC reference table}
\usage{
ABCRef(npart, priors, func, data, parallel = F, mc.cores = NA, ...)
}
\arguments{
\item{npart}{The number of particles (must be a positive integer).}

\item{priors}{A \code{data.frame} containing columns \code{parnames}, \code{dist}, \code{p1} and 
\code{p2}, with number of rows equal to the number of parameters. The column
\code{parname} simply gives names to each parameter for plotting and summarising.
Each entry in the \code{dist} column must contain one of \code{c("unif", "norm", "gamma")}, 
and the corresponding \code{p1} and \code{p2} entries relate to the hyperparameters 
(lower and upper bounds in the uniform case; mean and standard deviation in the 
normal case; and shape and rate in the gamma case).}

\item{func}{Function that runs the simulator. The first argument must be \code{pars}. The function
must return a \code{vector} of simulated summary measures, or a missing value (\code{NA})
if there is an error. The output from the function must be a vector with length equal 
to \code{nrow(data)} and with entries in the same order as the rows of \code{data}.}

\item{data}{A \code{data.frame} with a single row and columns containing the observed summary
statistics in order to extract output names.}

\item{parallel}{A \code{logical} determining whether to use parallel processing or not.}

\item{mc.cores}{Number of cores to use if using parallel processing.}

\item{...}{Extra arguments to be passed to \code{func}.}
}
\value{
An \code{data.frame} object with \code{npart} rows, where the first \code{p} columns correspond to 
        the proposed parameters, and the remaining columns correspond to the simulated outputs.
}
\description{
Runs ABC reference table
}
