% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABCSMC-package.R
\docType{package}
\name{ABCSMC-package}
\alias{ABCSMC-package}
\title{Sequential Monte Carlo based ABC}
\description{
Package implements the ABC-SMC algorithm of Toni et al. (2009)
}
\details{
Package provides some simple code to run the ABCSMC algorithm of Toni et al. (2009)
and plot the outputs.
}
\author{
TJ McKinley <t.mckinley@exeter.ac.uk>
}
