% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABCSMC.R
\name{ABCSMC}
\alias{ABCSMC}
\alias{ABCSMC.ABCSMC}
\alias{ABCSMC.default}
\title{Runs ABC-SMC algorithm}
\usage{
ABCSMC(x, ...)

\method{ABCSMC}{ABCSMC}(x, tols, mc.cores = 1)

\method{ABCSMC}{default}(npart, tols, priors, func, data, mc.cores = 1, ...)
}
\arguments{
\item{x}{An \code{ABCSMC} object.}

\item{tols}{A \code{matrix} of tolerances, with the number of rows defining
the number of generations required, and the number of columns
defining the number of summary statistics / data points to match to.}

\item{mc.cores}{Number of cores to use if using parallel processing.}

\item{npart}{The number of particles (must be a positive integer).}

\item{priors}{A \code{data.frame} containing columns \code{parname}, \code{dist}, \code{p1} and 
\code{p2}, with number of rows equal to the number of parameters. The column
\code{parname} simply gives names to each parameter for plotting and summarising.
Each entry in the \code{dist} column must contain one of \code{c("unif", "norm", "gamma")}, 
and the corresponding \code{p1} and \code{p2} entries relate to the hyperparameters 
(lower and upper bounds in the uniform case; mean and standard deviation in the 
normal case; and shape and rate in the gamma case).
two columns containing the lower and upper bounds respectively.}

\item{func}{A function taking a single argument \code{pars} that runs the simulator
and returns the simulated summary measures against which to compare.}

\item{data}{A \code{vector} containing the observed summary statistics to match to.}
}
\value{
An \code{ABCSMC} object, essentially a \code{list} containing:
\itemize{
\item{\code{pars}:}{ a \code{list} of \code{matrix} objects containing the accepted
                     particles. Each element of the list corresponds to a generation 
                     of ABC-SMC, with each matrix being of dimension 
                     \code{npart} x \code{npars};}
\item{\code{output}:}{ a \code{list} of \code{matrix} objects containing the simulated
                     summary statistics. Each element of the list corresponds to a
                     generation of ABC-SMC, with each matrix being of dimension 
                     \code{npart} x \code{length(data)};}
\item{\code{weights}:}{ a \code{list} of \code{vector} objects containing the particle
                     weights. Each element of the list corresponds to a
                     generation of ABC-SMC, with each vector being of length
                     \code{npart};}
\item{\code{accrate}:}{ a \code{vector} of length \code{nrow(tols)} containing the
                     acceptance rates for each generation of ABC;}
\item{\code{tols}:}{ a copy of the \code{tols} input;}
\item{\code{priors}:}{ a copy of the \code{priors} input;}
\item{\code{data}:}{ a copy of the \code{data} input;}
\item{\code{func}:}{ a copy of the \code{func} input.}
}
}
\description{
Runs ABC-SMC algorithm of Toni et al. (2009)
}
