% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileRcpp.R
\name{compileRcpp}
\alias{compileRcpp}
\title{Compiles \code{SimBIID_model} object}
\usage{
compileRcpp(model = NULL)
}
\arguments{
\item{model:}{An object of class \code{SimBIID_model}.}
}
\value{
An object of class \code{XPtr} that points to the compiled function, or
        an R \code{function} object for calling directly from R.
}
\description{
Compiles an object of class \code{SimBIID_model} into an
             \code{XPtr} object for use in Rcpp functions, or an
             object of class \code{function} for calling directly from R.
}
