% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimBIID-package.R
\docType{package}
\name{SimBIID-package}
\alias{SimBIID-package}
\title{Simulation-based inference for infectious disease models}
\description{
Package implements various simulation-based inference routines for infectious
disease models.
}
\details{
Provides some code to run simulations of state-space models, and then
}
\author{
Trevelyan J. McKinley <t.mckinley@exeter.ac.uk>
}
