% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimBIID-package.R
\docType{package}
\name{SimBIID-package}
\alias{SimBIID-package}
\title{Simulation-based inference for infectious disease models}
\description{
Package implements various simulation-based inference routines for infectious
disease models.
}
\details{
Package provides some code to run the ABC-SMC algorithm of Toni et al. (2009)
and the Alive Particle Filter based particle MCMC algorithm (Jasra et al., 2013;
Drovandi et al., 2016). Also provides functions to plot the outputs.
}
\author{
TJ McKinley <t.mckinley@exeter.ac.uk>
}
