% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryABC.R
\name{summary.ABCSMC}
\alias{summary.ABCSMC}
\title{Summarises \code{ABCSMC} objects}
\usage{
\method{summary}{ABCSMC}(object, gen = NA, transfunc = NA, ...)
}
\arguments{
\item{object}{An \code{ABCSMC} object.}

\item{gen}{The generation of ABC that you want to extract. If left missing then
defaults to final generation.}

\item{transfunc}{Is a \code{function} object where the arguments to the function must
match all or a subset of the parameters in the model. This function needs 
to return a \code{data.frame} object with columns containing the transformed
parameters.}

\item{...}{Not used here.}
}
\value{
A \code{data.frame} with weighted posterior means and variances.
}
\description{
Summary method for \code{ABCSMC} objects.
}
